+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## 🐞 Fixes
 - Fixed ovipositors suffocating.
+- Fixed facehugger lungs expanding even when not on a host.
 - Fixed gigeresque acid melting through chitin and resin blocks.
 - Fixed the mod interfering with enderman spawns in The End.
   - Xenomorphs were added to The End's spawn pool on the off chance that they spread to The End. However because Minecraft uses a weighted spawn system, xenomorphs being introduced decreased the chances of endermen spawning. To fix this issue, we've had to remove natural xenomorph spawns from The End's spawn pool.